refactor(index.ts): migrate typescript types to .ts files
Refactor project to move TypeScript type definitions from .d.ts files to
.ts files. Update import paths and references accordingly across the
codebase. Adjust theme and folder types to use more generic types where
necessary. Update constants imports to use new utils location. No
functional changes introduced.
